pyproject.toml
[tool.poetry]
name = "ssd_head_keras"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pillow = "^7.1.2"
opencv-python = "^4.2.0"
tensorflow = "~1.15.0"
keras = "==2.2.4"
numpy = "^1.18.4"
matplotlib = "^3.2.1"

[tool.poetry.dev-dependencies]
notebook = "^6.0.3"
pylint = "^2.5.2"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
